@@ -106,33 +104,37 @@
     "karma-sourcemap-loader": "^0.3.7",
     "karma-spec-reporter": "0.0.26",
     "karma-webpack": "^3.0.0",
-    "lolex": "^1.5.1",
-    "markdown-it": "^6.1.1",
-    "markdown-it-anchor": "^2.5.0",
+    "markdown-it": "^8.4.1",
+    "markdown-it-anchor": "^5.0.2",
     "markdown-it-container": "^2.0.0",
+    "mini-css-extract-plugin": "^0.4.1",
     "mocha": "^3.1.1",
     "node-sass": "^4.5.3",
+    "optimize-css-assets-webpack-plugin": "^5.0.1",
     "perspective.js": "^1.0.0",
     "postcss": "^5.1.2",
-    "postcss-loader": "0.11.1",
-    "postcss-salad": "^1.0.8",
+    "postcss-loader": "^2.1.5",
+    "postcss-salad": "^2.0.1",
     "progress-bar-webpack-plugin": "^1.11.0",
     "rimraf": "^2.5.4",
     "sass-loader": "^6.0.6",
+    "select-version-cli": "^0.0.2",
     "sinon": "^1.17.6",
     "sinon-chai": "^2.8.0",
     "style-loader": "^0.19.0",
     "transliteration": "^1.1.11",
+    "uglifyjs-webpack-plugin": "^2.1.1",
     "uppercamelcase": "^1.1.0",
-    "url-loader": "^0.6.2",
-    "vue": "^2.5.2",
-    "vue-loader": "^13.3.0",
-    "vue-markdown-loader": "1",
-    "vue-router": "2.7.0",
-    "vue-template-compiler": "^2.5.2",
+    "url-loader": "^1.0.1",
+    "vue": "^2.5.16",
+    "vue-loader": "^15.2.4",
+    "vue-markdown-loader": "^2.4.1",
+    "vue-router": "^3.0.1",
+    "vue-template-compiler": "^2.5.16",
     "vue-template-es2015-compiler": "^1.6.0",
-    "webpack": "^3.7.1",
-    "webpack-dev-server": "^2.9.1",
-    "webpack-node-externals": "^1.6.0"
+    "webpack": "^4.14.0",
+    "webpack-cli": "^3.0.8",
+    "webpack-dev-server": "^3.1.11",
+    "webpack-node-externals": "^1.7.2"
   }
 }
