@@ -69,6 +69,8 @@ import Header from '../packages/header/index.js';
 import Aside from '../packages/aside/index.js';
 import Main from '../packages/main/index.js';
 import Footer from '../packages/footer/index.js';
+import Timeline from '../packages/timeline/index.js';
+import TimelineItem from '../packages/timeline-item/index.js';
 import locale from 'element-ui/src/locale';
 import CollapseTransition from 'element-ui/src/transitions/collapse-transition';
 
@@ -138,6 +140,8 @@ const components = [
   Aside,
   Main,
   Footer,
+  Timeline,
+  TimelineItem,
   CollapseTransition
 ];
 
@@ -171,8 +175,8 @@ if (typeof window !== 'undefined' && window.Vue) {
   install(window.Vue);
 }
 
-module.exports = {
-  version: '2.4.11',
+export default {
+  version: '2.5.4',
   locale: locale.use,
   i18n: locale.i18n,
   install,
@@ -245,7 +249,7 @@ module.exports = {
   Header,
   Aside,
   Main,
-  Footer
+  Footer,
+  Timeline,
+  TimelineItem
 };
-
-module.exports.default = module.exports;
