@@ -1,5 +1,5 @@
-import navConfig from './nav.config.json';
-import langs from './i18n/route.json';
+import navConfig from './nav.config';
+import langs from './i18n/route';
 
 const LOAD_MAP = {
   'zh-CN': name => {
@@ -16,6 +16,11 @@ const LOAD_MAP = {
     return r => require.ensure([], () =>
       r(require(`./pages/es/${name}.vue`)),
     'es');
+  },
+  'fr-FR': name => {
+    return r => require.ensure([], () =>
+      r(require(`./pages/fr-FR/${name}.vue`)),
+    'fr-FR');
   }
 };
 
@@ -38,6 +43,11 @@ const LOAD_DOCS_MAP = {
     return r => require.ensure([], () =>
       r(require(`./docs/es${path}.md`)),
     'es');
+  },
+  'fr-FR': path => {
+    return r => require.ensure([], () =>
+      r(require(`./docs/fr-FR${path}.md`)),
+    'fr-FR');
   }
 };
 
@@ -146,6 +156,8 @@ if (userLanguage.indexOf('zh-') !== -1) {
   defaultPath = '/zh-CN';
 } else if (userLanguage.indexOf('es') !== -1) {
   defaultPath = '/es';
+} else if (userLanguage.indexOf('fr') !== -1) {
+  defaultPath = '/fr-FR';
 }
 
 route = route.concat([{
